prueba de Grafica

+++ app/src/main/java/molina/raul/aquareminder2/Consumo.kt
@@ -2,4 +2,4 @@ package molina.raul.aquareminder2
 
 import java.util.*
 
-data class Consumo(var cantidad: Int, var fecha: Date)
+data class Consumo(var cantidad: Double, var fecha: Date)
